Pietro/requirements.txt
anthropic>=0.18.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

